@@ -8,6 +8,7 @@
     "generate": "graphql-codegen --config codegen.yml",
     "test": "cross-env TEST=TRUE jest ",
     "drive-fetch": "ts-node --project tsconfig.server.json scripts/drive-fetch.ts",
+    "mailing-list": "ts-node --project tsconfig.server.json scripts/mailing-list.ts",
     "test-off": "ts-node --project tsconfig.server.json scripts/testOpenFoodFacts.ts",
     "db-setup": "ts-node --project tsconfig.server.json scripts/db-setup.ts",
     "all-notif": "ts-node --project tsconfig.server.json scripts/sendNotificationToAllCustomer.ts",
@@ -17,13 +18,17 @@
   },
   "dependencies": {
     "@apollo/client": "^3.2.2",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.30",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
     "@graphql-codegen/typescript-react-apollo": "^2.2.2",
-    "@material-ui/core": "^4.10.1",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@lottiefiles/react-lottie-player": "^3.4.1",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/lab": "^5.0.0-alpha.57",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.2.1",
     "@types/google-map-react": "^2.1.0",
     "@types/react-router-dom": "^5.1.7",
     "apollo-server-env": "^2.4.5",
@@ -37,13 +42,14 @@
     "isomorphic-unfetch": "^3.0.0",
     "js-cookie": "^2.2.1",
     "js-cookies": "^1.0.4",
+    "jsonwebtoken": "^8.5.1",
     "mysql": "^2.18.1",
-    "next": "latest",
+    "next": "^12.0.2",
     "pdfmake": "^0.2.2",
-    "react": "^16.8.4",
+    "react": "^17.0.2",
     "react-beautiful-dnd": "^13.0.0",
     "react-checkbox-list": "0.0.2",
-    "react-dom": "^16.8.4",
+    "react-dom": "^17.0.2",
     "react-i18next": "^11.5.0",
     "react-list-drag-and-drop": "^0.9.2",
     "react-movable": "^2.5.1",
@@ -59,12 +65,15 @@
     "@graphql-codegen/introspection": "1.13.1",
     "@graphql-codegen/typescript": "^1.13.1",
     "@graphql-codegen/typescript-resolvers": "1.13.1",
+    "@types/google-map-react": "^2.1.0",
     "@types/googlemaps": "^3.43.1",
     "@types/jest": "^26.0.9",
+    "@types/jsonwebtoken": "^8.5.6",
     "@types/mysql": "^2.15.13",
     "@types/node": "^12.0.12",
     "@types/react": "^16.8.17",
     "@types/react-dom": "16.8.4",
+    "@types/react-router-dom": "^5.1.7",
     "@types/sqlite3": "^3.1.6",
     "@types/styled-components": "^5.1.0",
     "apollo-server": "^2.16.1",
