@@ -1,45 +1,1 @@
-export type TranslationEnum =
-  | "hello.label"
-  | "thankyou.label"
-  | "darkmode.label"
-  | "lightmode.label"
-  | "changeLanguageButton.label"
-  | "buttonExample.label"
-  | "soto.description"
-  | "textfieldExample.label"
-  | "loginSoto.label"
-  | "baseline"
-  | "aboutUs"
-  | "privacyPolicy"
-  | "conditionsOfUse"
-  | "contactUs"
-  | "settings.personalInfos"
-  | "password.label"
-  | "newPassword.label"
-  | "confirmNewPassword.label"
-  | "button.criteria.label"
-  | "button.obligation.label"
-  | "searchbar.placeholder.label"
-  | "obligation.vegetarian"
-  | "obligation.vegan"
-  | "obligation.peanut_free"
-  | "obligation.lactose_free"
-  | "obligation.bio"
-  | "obligation.halal"
-  | "obligation.kosher"
-  | "obligation.gluten_free"
-  | "criteria.health"
-  | "criteria.environmental_impact"
-  | "criteria.price"
-  | "criteria.nutritional_value"
-  | "criteria.product_proximity"
-  | "shop.tooltip.label"
-  | "shop.banner.myBasket"
-  | "shop.banner.savedListButton.label"
-  | "shop.totalPrice.label"
-  | "shop.payButton.label"
-  | "error.notloggedin"
-  | "error.invalidcredentails"
-  | "error.emailalreadyinuse"
-  | "error.badparams";
-export const ErrMsg = (inp: TranslationEnum) => inp as string;
+export type TranslationEnum = "hello.label"|"thankyou.label"|"darkmode.label"|"lightmode.label"|"changeLanguageButton.label"|"buttonExample.label"|"soto.description"|"textfieldExample.label"|"loginSoto.label"|"baseline"|"aboutUs"|"privacyPolicy"|"conditionsOfUse"|"contactUs"|"settings.personalInfos"|"settings.personalization"|"language"|"email.label"|"password.label"|"newPassword.label"|"confirmNewPassword.label"|"button.criteria.label"|"button.obligation.label"|"searchbar.placeholder.label"|"obligation.vegetarian"|"obligation.vegan"|"obligation.peanut_free"|"obligation.lactose_free"|"obligation.halal"|"obligation.kosher"|"obligation.gluten_free"|"obligation.bio"|"criteria.health"|"criteria.environmental_impact"|"criteria.price"|"criteria.promotion"|"criteria.nutritional_value"|"criteria.product_proximity"|"shop.tooltip.label"|"shop.banner.myBasket"|"shop.banner.savedListButton.label"|"shop.totalPrice.label"|"shop.payButton.label"|"label.showMore"|"label.showLess"|"label.cart"|"label.saved_list"|"label.total"|"label.project"|"label.project.l1"|"label.project.l2"|"label.project.l3"|"label.startShopping"|"label.team"|"label.team.front"|"label.team.back"|"label.team.mobile"|"label.january.short"|"label.february.short"|"label.march.short"|"label.april.short"|"label.may.short"|"label.june.short"|"label.july.short"|"label.august.short"|"label.september.short"|"label.october.short"|"label.november.short"|"label.december.short"|"label.history"|"label.fullHistory"|"label.products_number"|"label.continue_shopping"|"label.history.cart"|"label.addToCart"|"label.boughtOn"|"label.at"|"label.history.paid"|"label.history.orderedProducts"|"label.quantity"|"label.price"|"label.score.final"|"title.profile"|"label.email"|"label.changeEmail"|"label.helperText.emailInvalid"|"label.password"|"label.changePassword"|"label.helperText.passwordsDifferent"|"label.language"|"label.usedLanguage"|"label.accessibility"|"label.lightTheme"|"label.darkTheme"|"label.changeTheme"|"label.themeSelection.helperText"|"label.criteriaAndObligations"|"description.criteriaAndObligations"|"label.driveSelection"|"description.driveSelection"|"language.french"|"language.location.fr"|"language.english"|"language.location.us"|"language.location.uk"|"label.general.cancel"|"label.general.validate"|"label.general.back"|"label.general.finish"|"label.general.next"|"label.confirmPassword"|"label.login"|"label.login.connect"|"label.login.signUp"|"label.login.registerToLogin"|"label.login.step1.title"|"label.login.step2.title"|"label.login.step3.title"|"notification.label.registered"|"notification.label.loggedIn"|"error.notloggedin"|"error.invalidcredentails"|"error.emailalreadyinuse"|"error.badparams"|"label.listName"|"label.milestones"|"label.milestones.step1"|"label.milestones.step2"|"label.milestones.step3"|"label.milestones.step4"|"label.milestones.step5.date"|"label.milestones.step5"|"label.copyEmail"|"notification.copyEmail"|"criteria.lowCalories"|"criteria.highProtein"|"obligation.no_lactose"; export const ErrMsg = (inp : TranslationEnum) => inp as string
