@@ -5,7 +5,7 @@ import { getPool } from "../server/query";
 const mysql = require("mysql");
 
 const insertProduct =
-  "(id INT AUTO_INCREMENT primary key NOT NULL, leclercId VARCHAR(64), photo VARCHAR(1024), name VARCHAR(1024), brand VARCHAR(1024), priceUnit INT, priceMass VARCHAR(64), ingredients VARCHAR(4096), packaging VARCHAR(4096), allergens VARCHAR(4096), nutriments VARCHAR(4096), nutriscore VARCHAR(2), healthscore INT, environmentScore INT, priceScore INT, proximityScore INT, quantity VARCHAR(16), keywords VARCHAR(4096), vegan BOOLEAN, noGluten BOOLEAN, labels VARCHAR(4096), bio BOOLEAN, peanutFree BOOLEAN)";
+  "(id INT AUTO_INCREMENT primary key NOT NULL, leclercId VARCHAR(64), photo VARCHAR(1024), name VARCHAR(1024), brand VARCHAR(1024), priceUnit VARCHAR(64), priceMass VARCHAR(64), ingredients VARCHAR(4096), packaging VARCHAR(4096), allergens VARCHAR(4096), nutriments VARCHAR(4096), nutriscore VARCHAR(2), healthscore INT, environmentScore INT, priceScore INT, promotionScore INT, proximityScore INT, quantity VARCHAR(16), keywords VARCHAR(4096), vegan BOOLEAN, noGluten BOOLEAN, labels VARCHAR(4096), bio BOOLEAN, peanutFree BOOLEAN, promotion VARCHAR(64), noLactose BOOLEAN, highProteinScore INT, lowCaloriesScore INT)";
 const sqlconnect = async () => {
   return new Promise<any>((resolve, reject) => {
     const con = mysql.createConnection(getPool());
@@ -40,7 +40,7 @@ const start = async () => {
 
   await sqlquery(
     sql,
-    "CREATE TABLE users (id INT AUTO_INCREMENT primary key NOT NULL, token VARCHAR(64), email VARCHAR(255) NOT NULL, password VARCHAR(64) NOT NULL, cartId INT, shopId INT, pushToken VARCHAR(64));"
+    "CREATE TABLE users (id INT AUTO_INCREMENT primary key NOT NULL, email VARCHAR(255) NOT NULL, password VARCHAR(64) NOT NULL, cartId INT, shopId INT, pushToken VARCHAR(64), mailingList BOOLEAN NOT NULL DEFAULT FALSE);"
   );
   await sqlquery(
     sql,
