@@ -12,6 +12,7 @@ export interface DbProduct {
   nutriments: string;
   nutriscore: string;
   healthscore: number;
+  promotionScore: number;
   environmentScore: number;
   priceScore: number;
   proximityScore: number;
@@ -20,11 +21,11 @@ export interface DbProduct {
   origin?: string;
   labels: string;
   ecoscore: number;
+  promotion: string;
 }
 
 export interface DbUser {
   id: number;
-  token: string;
   email: string;
   password: string;
   cartId: number;
